+++ db/schema.rb
@@ -13,6 +13,7 @@
 ActiveRecord::Schema.define(version: 2019_11_03_121005) do
 
   # These are extensions that must be enabled in order to support this database
+  enable_extension "pg_trgm"
   enable_extension "plpgsql"
 
   create_table "character_classes", force: :cascade do |t|