+++ frontend/packs/static_management.jsx
@@ -1,0 +1,4 @@
+import StaticManagement from "components_react/static_management/static_management";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({StaticManagement});